+++ .eslintrc.json
@@ -5,6 +5,7 @@
   ],
   "rules": {
     "react/prop-types": "off",
-    "no-unused-vars": "warn"
+    "no-unused-vars": "warn",
+    "no-undef": "off"
   }
 }
