+++ package.json
@@ -13,10 +13,13 @@
     "@auth0/nextjs-auth0": "^3.5.0",
     "@prisma/client": "^5.11.0",
     "@vercel/postgres": "^0.8.0",
+    "axios": "^1.7.2",
     "bcryptjs": "^2.4.3",
+    "cors": "^2.8.5",
     "date-fns": "^3.6.0",
     "i": "^0.3.7",
     "install": "^0.13.0",
+    "multer": "^1.4.5-lts.1",
     "next": "14.0.4",
     "next-auth": "^5.0.0-beta.16",
     "npm": "^10.2.5",
